fix: keep shipping fees in decimal form

+++ shopcore/shipping.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
+from decimal import Decimal
 
 from .utils import money
 
@@ -8,7 +9,7 @@ from .utils import money
 @dataclass(frozen=True)
 class ShippingQuote:
     method: str
-    fee: float
+    fee: Decimal
     eta_days: int
 
 
@@ -16,7 +17,7 @@ class ShippingService:
     def quote(self, shipping_country: str, item_count: int) -> ShippingQuote:
         if shipping_country == "US":
             if item_count > 3:
-                return ShippingQuote(method="ground", fee=float(money("7.50")), eta_days=3)
-            return ShippingQuote(method="standard", fee=float(money("5.00")), eta_days=5)
+                return ShippingQuote(method="ground", fee=money("7.50"), eta_days=3)
+            return ShippingQuote(method="standard", fee=money("5.00"), eta_days=5)
 
-        return ShippingQuote(method="international", fee=float(money("15.00")), eta_days=10)
+        return ShippingQuote(method="international", fee=money("15.00"), eta_days=10)
